+++ Packages/com.datea83.choyutilities/Runtime/Entities/Agents/AgentMoveIEnableable.cs
@@ -12,26 +12,33 @@ namespace EugeneC.ECS {
         public Entity CurrentNode;
         public float CurrentRestTime;
     }
+    
+    public struct AgentMoveICleanupTag : ICleanupComponentData{ }
 
-    public struct AgentMoveNodeIBuffer : IBufferElementData {
+    public struct ConnectedNodeIBuffer : IBufferElementData {
         
         public Entity ConnectedNode;
         
     }
 
-    public struct AgentSpawnIBuffer : IBufferElementData {
+    public struct SpawnNodeIBuffer : IBufferElementData {
 
         public Entity Prefab;
     }
+
+    public struct SpawnNodeTagIData : IComponentData {
+
+	    public byte Id;
+    }
     
-    public struct AgentSpawnNodeIData : IComponentData {
+    public struct SpawnNodeIEnableable : IComponentData, IEnableableComponent {
         
         public bool SpawnOnce;
         public float DefaultSpawnDelay;
         public float CurrentSpawnDelay;
     }
     
-    public struct AgentSpawnISingleton : IComponentData {
+    public struct AgentISingleton : IComponentData {
         
         public ushort TotalSpawnCount;
         public ushort CurrentSpawnCount;